Support for specifying headers to forward
In order to route custom domains to apps we offer a CDN using AWS
CloudFront. By default CloudFront only forwards a limited set of
HTTP headers, but our users have asked to forward more headers to
their apps (for instance the `Accept` header to indicate the
`Content-Type` to return.)

This commit adds a `headers` field to the provisioning details, which can
optionally be given as an array of strings. For instance:

 - {…, "headers": []} or omitting `headers` altogether will forward the
   `Host` header and Amazon's default headers.

 - {…, "headers": ["Accept"]} will forward the `Accept` header to the app,
   as well as `Host` and Amazon's default headers.

 - {…, "headers": ["*"]} will forward all headers and Amazon disables
   caching.

Amazon caps the number of whitelisted headers (as in the `Accept` example
above) to 10. We use `Host` of this 10 so that the CloudFoundry router
can map the custom domain to the app. This leaves 9 for our users, and if
they require more they can use the `*` wildcard header above.

CloudFront's caching mechanism maps forwarded HTTP requests to their
response. Only forwarding a limited set of headers by default maximises
the effectiveness of its cache. Therefore the more headers one forwards
(and the more variability within them) the less effective its caching
will be. We intend to warn about this in the documentation.

This commit also changes the type used for forwarded headers. Previously
they were passed around as a slice of strings, but our changes needed to
work with canonicalised headers and check for duplicates. It proved
cleaner and more natural to pass around a new `Headers` type, which
stores canonicalised headers and automatically deduplicates. This is
only converted to a `[]string` inside the CloudFront API utility.
`Headers` is defined in the `utils` package to avoid circular dependency.