Don't log the broker's settings during canary check
The broker's setting struct contains several secrets we'd prefer not to
log during domain/route liveness checks. We could blacklist or whitelist
specific fields in the logging call, but the additional value of this
information during a canary check is felt to be minimal, given that we have
visibility of the route/domain being requested, as we log it elsewhere if we
encounter a failure.

Instead, we're choosing to remove the settings struct entirely from the log.