Remove Georgetown DNS nameservers from tests
We now include part of an ACME library in our codebase. One of its
tests failed because it got different nameservers back from a domain
that the tests expected:

    dns_challenge_test.go:143: #physics.georgetown.edu.: got [ns4.georgetown.edu. ns5.georgetown.edu. ns6.georgetown.edu.]; want [ns1.georgetown.edu. ns2.georgetown.edu. ns3.georgetown.edu.]

Since there are other, working testcases it seems easiest to drop
Georgetown rather than alter the implementation to be more flexible.