Add info logging to Broker API endpoints
This commit adds info-level logging to the Open Service Broker API
endpoints. This logging will give us more context on the activity
happening in production.

This logging is being added on a "best-effort" basis. A larger
refactor/reabstraction seems to be needed, at which point more
thorough logging could be easily done.