Do not bother doing HTTP challenges in Update
When we are updating the domains, we always need to do a new DNS01
challenge.

Conversely, when we are not updating the domains, we NEVER need to do an
HTTP01 challenge as this is handled during Renew.

Therefore we can skip the HTTP challenge performed in Update, which
will make updating the normal configuration fast and simple.

Co-authored-by: Andy <andy.hunt@digital.cabinet-office.gov.uk>
Co-authored-by: Toby <toby.lornewelch-richards@digital.cabinet-office.gov.uk>
Signed-off-by: Toby Lorne <toby.lornewelch-richards@digital.cabinet-office.gov.uk>