Create accounts using the ACME V2 API
The V1 API used by the Lego client is no longer available, which means
people can't create new services anymore.

Manual testing indicates that the V1 API calls still work, even with a
V2 account, so we think the simplest fix is just to update the code to
create accounts to use the newer API.

It doesn't look like this is covered by tests yet, which isn't great.