Fill out the legoacme.Registration struct
When wiring the ACME v1 and v2 APIs together, we have to manually emulate the
response from the v1 API using v2 data. Previously, we left the `Body` field
blank because we thoguht it was unencessary. Later evidence shows we need it,
and this commit fills it out as best we can.