Check if a V2 ACME account exists before creating a new one
Previous versions of the golang.org/x/crypto/acme library returned an account
struct from `client.Register` if the account already existed. This version does
not. Instead, we must try fetch the registration, and if a particular errors is
returned, we must register another account. Otherwise, we use the account
returned by `client.GetReg`.

In particular, we should note that on `client.GetReg`, the second argument
(`url`) is ignored when the CA's API is RFC 8555 compliant (i.e. it's an
implementation of ACME v2)