Increase cdn-route response timeouts to 60 seconds
Until now our cdn-route service has had a request timeout of 30
seconds. Any user request which took longer than 30 seconds would
fail. This doesn't work well for a minority of tenants.

In this commit we configure the Origin Read Timeout [1] for the
first time. It defaults to 30 seconds and can be set between 4–60
seconds. We set it to 60 to give requests the best chance of
succeeding. We previously did that manually for one tenant so we're
fairly confident it has no major surprises for us.

This repo lacks tests around the `Distribution` and
`CustomOriginConfig` types and that's much too big to try and fix now.
As such this has no automated tests and we'll need to test it against
CloudFront.

[1] https://aws.amazon.com/about-aws/whats-new/2017/03/announcing-configure-read-timeout-and-keep-alive-timeout-values-for-your-amazon-cloudfront-custom-origins/