Upgrade AWS SDK to support OriginReadTimeout
The previous commit starts setting the field
`CustomOriginConfig.OriginReadTimeout`, used to tell CloudFront how
long to wait for an upstream response before timing out. The version
of the AWS SDK needed updating to support that field.

AWS has new API versions which are used by the upgraded AWS SDK, so
this commit also fixes the mocking.

This project is not massively tested so we'll have to use it and build
confidence in a dev environment.