UpdateOptions should be pointers
If you have a struct like

type UpdateOptions struct {
	Domain     string   `json:"domain"`
	DefaultTTL int64    `json:"default_ttl"`
	Cookies    bool     `json:"cookies"`
	Headers    []string `json:"headers"`
}

And you unmarshal '{}' into it

then the struct will have default values of "" for a string, 0 for an
int64, [] for an array, etc

If you have a struct like

type UpdateOptions struct {
	Domain     *string   `json:"domain,omitempty"`
	DefaultTTL *int64    `json:"default_ttl,omitempty"`
	Cookies    *bool     `json:"cookies,omitempty"`
	Headers    *[]string `json:"headers,omitempty"`
}

And you unmarshal {}' into it

then the struct will have nil values for *string, *int64, *[] etc

This is important. Currently if you

$ cf update-service my-cdn-route -c '{"domain": "my.domain.com", "headers": ["Accept", "Authorization"]}'

Then the headers will be set correctly

Then later, you do

$ cf update-service my-cdn-route -c '{"domain": "my.domain.com,my.otherdomain.com"}'

The resultant CDN distribution will have the headers reset

Now this not the case because we can distinguish between an empty list
and a not-provided list. This can be formalised as the difference
between Bool and Maybe<Bool>, where nil or not functions as the Maybe
monad.

Future commits update the type signature and mocks and tests

Signed-off-by: Toby Lorne <toby.lornewelch-richards@digital.cabinet-office.gov.uk>