Do not disable and delete CloudFront distribution on timeout
Previously, when the 84h provisioning window had passed we disabled the
CloudFront disribution and deleted it in a subsequent run of the background
job.

This was intended to make it easier for new distributions to be provisioned
without CloudFront's globally unique restriction on CNAMEs being hit. However,
we've discovered a bug (a potential race condition) which causes the
`ProvisioningSince` time to be set to `null` when provisioning (or updating)
has finished. As a result, distributions were being deleted at the beginning of
a new timeout period.

This is not good.

We should not do that. We should instead mark the service as failed and allow
tenants to delete it instead.