Log, but ignore, missing ACM certs when fetching DNS challenges
Inspection of the data in the production CDN broker database shows that in some
cases we're storing the ARN of a certificate which no longer exists. We don't
know why this is, or exactly how many there are. However, if they don't exist
then they can't be pending validation 🧐, so it's safe to log the error for
posterity, and carry on to fetch the rest of the certificates.