config: add DatabaseConnMaxLifetime & DatabaseMaxIdleConns settings
long lived idle connections seem to be able to timeout in a way that
the connection pool fails to detect. setting DatabaseConnMaxLifetime
should hopefully reduce the degree to which that happens.

keeping the DatabaseMaxIdleConns at 1 should reduce the possibility
that healthchecks will overlook broken connections sitting at the
back of the pool

pull github.com/pkg/errors in as a direct dependency as it's too
useful to not be able to use